[MOV] mail: move suggested recipients on BaseModel
We plan to use suggested recipients in templates, which means it should
be available as a generic methods instead of limited to mail.thread.
BaseModel already has some mail-related methods due to templates:
finding customers, default recipients, ... we now add suggested
recipients for the same reasons.

Task-4555506: Cleanup / Mergeup default / suggested recipients
